Thining out controller
I have moved most of the logic and functionality
out into the models for the user.rb. I have created
some accessor class methods that allow me to query
for user objects. I have overridden the 'serializable_hash'
method of user so that it sends down the appropriate data
automatically.